+++ src/GraphiQL/AuthenticationFailedException.php
@@ -10,10 +10,18 @@
 
 namespace Ynlo\GraphQLBundle\GraphiQL;
 
+use Throwable;
+
 /**
  * AuthenticationFailedException
  */
 class AuthenticationFailedException extends \Exception
 {
-
+    /**
+     * @inheritDoc
+     */
+    public function __construct(string $message = 'Authentication Failed', int $code = 0, Throwable $previous = null)
+    {
+        parent::__construct($message, $code, $previous);
+    }
 }